% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_luxembourg.R
\docType{data}
\name{data_luxembourg}
\alias{data_luxembourg}
\alias{ver1_0_LU_1km_pt_ppl_within_10-20_min.csv.gz}
\title{Example accessibility CSV for Luxembourg}
\format{
A gzipped CSV file with 1 600 rows and 3 columns:
\describe{
\item{lat}{\emph{dbl}; latitude (decimal degrees)}
\item{lon}{\emph{dbl}; longitude (decimal degrees)}
\item{opportunities (people)}{\emph{dbl}; number of people reachable in 10–20 min on public transport}
}
}
\source{
Transport Poverty Hub © European Union, 1995–2025.
The Commission’s reuse policy is implemented by the
\href{https://eur-lex.europa.eu/legal-content/EN/TXT/?uri=CELEX:32011D0833}{Commission Decision of 12 December 2011 on the reuse of Commission documents}. Unless otherwise indicated, content owned by the EU on this website
is licensed under the \href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution 4.0 International (CC BY 4.0) licence}. Reuse is allowed, provided appropriate credit is given and changes are indicated.

More information is available at:
\itemize{
\item \href{https://urban-mobility-observatory.transport.ec.europa.eu/news-events/news/recommendations-member-states-help-tackle-transport-poverty-and-promote-fair-sustainable-mobility-2025-05-22_en}{Recommendations for Member States to help tackle transport poverty and promote fair, sustainable mobility}
\item \href{https://jeodpp.jrc.ec.europa.eu/eu/vaas/voila/render/mobitrans/transportpoverty/TransportPovertyHub.ipynb}{Transport Poverty Hub (requires login with a free EU account)}
}
}
\description{
A small gzipped CSV with access to opportunities measured as the
number of people reachable within 10–20 minutes on public transport
on a 1 km grid in Luxembourg.
}
\section{File location}{

The raw CSV is installed at:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{system.file(
  "extdata",
  "ver1_0_LU_1km_pt_ppl_within_10-20_min.csv.gz",
  package = "tphconv"
)
}\if{html}{\out{</div>}}
}

\examples{
f <- system.file(
  "extdata",
  "ver1_0_LU_1km_pt_ppl_within_10-20_min.csv.gz",
  package = "tphconv"
)
library(readr)
acc <- read_csv(f)
head(acc)

}
\keyword{datasets}
\keyword{internal}
