% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_grid_id_from_coords}
\alias{generate_grid_id_from_coords}
\title{Generate INSPIRE / GISCO Grid-Cell IDs from Coordinates}
\usage{
generate_grid_id_from_coords(xll, yll, cellsize, epsg, style, origin = c(0, 0))
}
\arguments{
\item{xll}{A numeric vector of x-coordinates for the lower-left (south-west)
corner of each cell.}

\item{yll}{A numeric vector of y-coordinates for the lower-left (south-west)
corner of each cell.}

\item{cellsize}{The grid resolution in the units of the coordinates.}

\item{epsg}{The integer EPSG code of the coordinate reference system.}

\item{style}{The format of the output ID string: "gisco_full" or "gisco_short".}

\item{origin}{A numeric vector of \code{c(x, y)} coordinates for the grid's origin.}
}
\value{
A character vector of grid-cell IDs.
}
\description{
A low-level helper function that computes standardized grid-cell identifiers
from numeric coordinate vectors.
}
\keyword{internal}
